requirements.txt
click>=8.0.0
rich>=13.0.0
requests>=2.28.0
aiohttp>=3.8.0
dnspython>=2.3.0
python-whois>=0.8.0
paramiko>=3.0.0
beautifulsoup4>=4.12.0
scapy>=2.5.0
reportlab>=4.0.0
urllib3>=1.26.0
